remove duplicate files
