+++ imports/containers/Component/AccountsWrapper/index.js
@@ -1,6 +1,4 @@
 import React, { Component } from "react";
-import ReactDOM from "react-dom";
-import { Accounts } from 'meteor/accounts-base'
 import { Form, Field } from 'react-final-form';
 import Button from '@material-ui/core/Button';
 import FormControl from '@material-ui/core/FormControl';
@@ -12,10 +10,9 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Radio from '@material-ui/core/Radio';
 import RadioGroup from '@material-ui/core/RadioGroup';
-// import FormHelperText from '@material-ui/core/FormHelperText';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
-// import FormControl from '@material-ui/core/FormControl';
 import FormLabel from '@material-ui/core/FormLabel';
+import { Meteor } from "meteor/meteor"
 
 export default class AccountsUIWrapper extends Component {
   constructor(props) {
@@ -23,62 +20,39 @@ export default class AccountsUIWrapper extends Component {
     this.state = {
       formToggle: true,
       error: null,
-      type: 'customer'
+      usertype: ''
     };
   }
-  signUp(user) {
+  signUp = (user) => {
     const { email, password, usertype } = user;
-    Accounts.createUser(
-      {
-        "email": email, "password": password
-      }, function (error) {
+    Meteor.call("users.createUser", email, password, usertype, (err, res) => {
+      if (err) console.log(err)
+      else console.log(res)
+    })
 
-        if (Meteor.user()) {
-          console.log("User created: ", Meteor.userId());
-          //add the usertype to the database
-        } else {
-          console.log("ERROR: " + error.reason);
-        }
-      }
-    );
-
-    Accounts.onCreateUser(function (options, user) {
-      // Use provided profile in options, or create an empty object
-      user.profile = options.profile || {};
-      // Assigns first and last names to the newly created user object
-      user.profile.type = usertype;
-      // Returns the user object
-      return user;
-    });
   }
 
-  logIn(user) {
+  logIn = (user) => {
     const { email, password } = user;
-    Meteor.loginWithPassword({ email, password }, function (error) {
-
-      if (Meteor.user()) {
-        console.log(Meteor.userId());
-      } else {
-        console.log("ERROR: " + error.reason);
-      }
-    });
+    Meteor.call("users.logIn", email, password)
   }
 
-  handleChange(event) {
-    this.setState({ type: event.target.value })
+  handleChange = (event) => {
+    this.setState({ usertype: event.target.value })
   }
 
   render() {
     const { classes } = this.props;
     return (
       <Form
-        // validate={validate.bind(this)}
+        // TODO validate={validate.bind(this)}
         onSubmit={values => {
-          const user = { email: values.email, password: values.password, type: values.usertype };
+          const user = { email: values.email, password: values.password, usertype: this.state.usertype };
           this.state.formToggle
             ? this.logIn(user)
             : this.signUp(user);
-          console.log(values);
+          console.log(user);
+          console.log("usertype is: ", this.state.usertype)
         }}
         render={({ handleSubmit, pristine, invalid, form }) => (
           <form onSubmit={handleSubmit}
@@ -93,14 +67,14 @@ export default class AccountsUIWrapper extends Component {
                   {({ input, meta }) => {
                     return (
                       <RadioGroup
-                        aria-label="gender"
-                        name="gender1"
+                        aria-label="usertype"
+                        name="usertype"
                         // className={classes.group}
                         value={input.value}
-                        onChange={this.handleChange.bind(this)}
+                        onChange={this.handleChange}
                       >
-                        <FormControlLabel value="restaurant" checked={this.state.type == 'restaurant'} control={<Radio />} label="Restaurant" />
-                        <FormControlLabel value="customer" checked={this.state.type == 'customer'} control={<Radio />} label="Customer" />
+                        <FormControlLabel value="restaurant" checked={this.state.usertype == 'restaurant'} control={<Radio />} label="Restaurant" />
+                        <FormControlLabel value="customer" checked={this.state.usertype == 'customer'} control={<Radio />} label="Customer" />
                       </RadioGroup>
                     )
                   }}
@@ -174,7 +148,6 @@ export default class AccountsUIWrapper extends Component {
                     // className={classes.formToggle}
                     type="button"
                     onClick={() => {
-                      // form.reset();
                       this.setState({
                         formToggle: !this.state.formToggle
                       });
