+++ package.json
@@ -17,6 +17,7 @@
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-final-form": "^6.3.0",
+    "react-gravatar": "^2.6.3",
     "react-router-dom": "^5.0.1"
   },
   "meteor": {
